test: add failing test of extensions yaml config
Trying to run `rst2myst convert --config` with a YAML config file
approximating the one in the documentation, even with the `--dry-run`
flag, causes `rst2myst` to emit output including a message indicating
a failure to import extensions. This commit implements a test that
confirms this behavior so that a fix can be prepared.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>