fix: convert stringified extension list to list
The default option type in `click` is a string, so `click` converts a
list from a YAML configuration file into a string -- essentially, the
Python-equivalent list `["sphinx-design"]` from a YAML input file is
passed to the `rst_to_myst.cli.split_extension` method's `value`
argument as the Python string `'["sphinx-design"]'`. This parameter
value then fouls up later logic in the
`rst_to_myst.namespace.compile_namespace` method, specifically the
`for extname in chain(builtin_extensions, extensions):` loop.

One way to recover the intended behavior is to use Python's `ast`
module to interpret the string as a fragment of Python source code,
which would then convert the stringified list into a Python list. This
commit makes that change. Admittedly, it's a little bit of a hack --
and uses a crude heuristic to decide when to attempt to convert the
string to a Python list -- but it seems like a good minimal change to
submit as part of a pull request.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>