fix: convert stringified conversions map to map
As with the extensions list, a pathology occurs when reading
conversions settings from an `rst2myst` YAML configuration file that
is similar to the bug exhibited when trying to read a list of
extensions from an `rst2myst` YAML configuration file: the conversions
map in the YAML configuration file is interpreted as a string, instead
of as a dictionary or a mapping.

In this case, the stringified map from the YAML configuration file is
passed into the `rst_to_myst.cli.read_conversions` method as the value
of its `value` parameter. Execution passes to the to the `if not
path.exists():` conditional in that method, which interprets the
string as a path, and checks to see if it exists. Although POSIX would
likely permit users to create a file with that string as a name, in
general, users probably would not name a file with braces.

The fix in this commit is probably not as robust as it could be. A
better solution would probably be to test if the path exists, and if
it doesn't exist, then attempt to convert the string to a map using
Python's `ast` module. Nevertheless, the fix in this commit is similar
to the fix for the extensions list: convert the string to a map using
the Python `ast` module from the Python standard library.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>