test: test for conversions & absence of failure
To exercise the changes to the `rst_to_myst.cli.read_conversions`
method, this commit adds a `conversions:` key to the YAML
configuration file used in the test. In addition, the test now checks
to see that the string `"FAILURE"` (no quotes) is absent from the
output, instead of present.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>