fix: add docs extra to CI tests
The `sphinx_design` module is needed to run the new YAML configuration
file test because that test needs a Sphinx plugin to exercise the code
that triggered the bugs motivating the changes. This commit adds the
`docs` extra to the `flit` install command so `flit` also installs the
`sphinx_design` module, which should resolve the CI failures.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>