fix: split_extensions: check if value is string
The `rst_to_myst.cli.split_extensions` callback will fail if `value`
is `None` because the statement assigning `could_be_list` a value will
fail if `value` is `None`.

This change is probably overkill, but it guards that assignment and
the `ast` module method call with a test on whether `value` is a
string. A simpler fix would be to test whether `value` is `None`, but
I haven't worked on this project long, so better safe than sorry.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>