maintain: github-actions via dependabot.yml
Some of the GitHub Actions used in the project CI are deprecated and
slated for end-of-life. To help automate the process of keeping the
actions used by this project current, this commit adds a
`dependabot.yml` file that will automatically file pull requests
weekly with suggested updates for GitHub Actions.

Signed-off-by: Geoffrey M Oxberry <goxberry@gmail.com>